dio/fastapi-azure-functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

fastapi==0.100.0
uvicorn==0.22.0
azure-functions==1.13.2 
